Refactor Divide Code of mainWindow
